# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Minimarket API",
    "description": "Contrato OpenAPI para la gestion de productos, carritos y recursos principales del minimarket.",
    "contact": {
      "name": "Maamartinezr",
      "url": "https://github.com/Maamartinezr/Backend-II_Sumativa_S7_Grupo5-"
    },
    "license": {
      "name": "Sin licencia definida",
      "url": "https://github.com/Maamartinezr/Backend-II_Sumativa_S7_Grupo5-"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Entorno local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener producto por id",
        "description": "Recupera el detalle de un producto por su identificador. Requiere autenticacion JWT.",
        "operationId": "obtenerProductoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                },
                "example": {
                  "id": 1,
                  "nombre": "Arroz integral",
                  "precio": 1500.0,
                  "stock": 20,
                  "categoriaId": 1
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Full authentication is required to access this resource",
                  "path": "/api/productos"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 403,
                  "error": "Forbidden",
                  "message": "Access Denied",
                  "path": "/api/productos"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Producto no encontrado con id: 99",
                  "path": "/api/productos/99"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Ha ocurrido un error inesperado",
                  "path": "/api/productos"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Productos"
        ],
        "summary": "Actualizar producto",
        "description": "Actualiza un producto existente. Requiere rol ADMIN.",
        "operationId": "actualizarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Datos del producto a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              },
              "example": {
                "nombre": "Arroz integral",
                "precio": 1500.0,
                "stock": 20,
                "categoriaId": 1
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                },
                "example": {
                  "id": 1,
                  "nombre": "Arroz integral",
                  "precio": 1500.0,
                  "stock": 20,
                  "categoriaId": 1
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 400,
                  "error": "Bad Request",
                  "message": "La solicitud contiene datos invalidos",
                  "path": "/api/productos"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Full authentication is required to access this resource",
                  "path": "/api/productos"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 403,
                  "error": "Forbidden",
                  "message": "Access Denied",
                  "path": "/api/productos"
                }
              }
            }
          },
          "404": {
            "description": "Producto o categoria no encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Producto no encontrado con id: 99",
                  "path": "/api/productos/99"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de datos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 409,
                  "error": "Conflict",
                  "message": "Conflicto de datos: el recurso ya existe o viola restricciones",
                  "path": "/api/productos"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Ha ocurrido un error inesperado",
                  "path": "/api/productos"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "summary": "Eliminar producto",
        "description": "Elimina un producto existente. Requiere rol ADMIN.",
        "operationId": "eliminarProducto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Producto eliminado correctamente"
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 401,
                  "error": "Unauthorized",
                  "message": "Full authentication is required to access this resource",
                  "path": "/api/productos"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 403,
                  "error": "Forbidden",
                  "message": "Access Denied",
                  "path": "/api/productos"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 404,
                  "error": "Not Found",
                  "message": "Producto no encontrado con id: 99",
                  "path": "/api/productos/99"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "timestamp": "2026-07-06T04:12:49.271Z",
                  "status": 500,
                  "error": "Internal Server Error",
                  "message": "Ha ocurrido un error inesperado",
                  "path": "/api/productos"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "example": "Arroz integral"
          },
          "precio": {
            "type": "number",
            "format": "double",
            "example": 1500.0
          },
          "stock": {
            "type": "integer",
            "format": "int32",
            "example": 20
          },
          "categoriaId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        },
        "required": [
          "categoriaId",
          "nombre",
          "precio",
          "stock"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-06T04:12:49.271Z"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "La solicitud contiene datos invalidos"
          },
          "path": {
            "type": "string",
            "example": "/api/productos"
          },
          "validationErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}